refactor(ale_claw): group durable-memory modules into harness/memory/ subpackage
Move memory, memory_flush, memory_flush_policy into a memory/ subpackage.
Behavior-preserving: git mv + import-path updates. 928 passed, 3 skipped.
