+++ ale_run/agents/ale_claw/harness/agent_loop.py
@@ -65,8 +65,8 @@ from ._message_shapes import _function_call_output, _image_url_block
 from .canonical.canonical import normalize_to_canonical, sanitize_items
 from .computer_handler import OpenClawComputerHandler
 from .context.context import ContextOverflowCallback, compact_messages, is_context_overflow_error
-from .memory import MemoryStore
-from .memory_flush import run_memory_flush
+from .memory.memory import MemoryStore
+from .memory.memory_flush import run_memory_flush
 from .prompt import ContextFile
 from .session import (
     MEMORY_FLUSH_PROMPT,
