+++ ale_run/agents/ale_claw/harness/memory/memory_flush.py
@@ -18,14 +18,14 @@ from __future__ import annotations
 import json as _json
 from typing import TYPE_CHECKING, Any
 
-from .canonical.canonical import _normalize_actions
-from .model_config import ResolvedModel, resolve_model
+from ..canonical.canonical import _normalize_actions
+from ..model_config import ResolvedModel, resolve_model
 
-from .helper_runtime import call_helper_model
+from ..helper_runtime import call_helper_model
 
 if TYPE_CHECKING:
     from .memory import MemoryStore
-    from .session import SessionManager
+    from ..session import SessionManager
 
 
 async def run_memory_flush(
