+++ tests/ale_claw/test_openclaw_memory_store.py
@@ -1,4 +1,4 @@
-"""Tests for ale_run.agents.ale_claw.harness.memory (MemoryStore + SearchResult).
+"""Tests for ale_run.agents.ale_claw.harness.memory.memory (MemoryStore + SearchResult).
 
 Ported from tests/test_memory_store.py, adapted for the task workspace layout
 where session logs live in a memory/ subdirectory under the task dir.
