+++ tests/ale_claw/test_openclaw_memory_tools.py
@@ -6,7 +6,7 @@ so CUA agents can search, read, and write memory files during task execution.
 
 import pytest
 
-from ale_run.agents.ale_claw.harness.memory import (
+from ale_run.agents.ale_claw.harness.memory.memory import (
     MemoryGetTool,
     MemorySearchTool,
     MemoryStore,
